requirements.txt
django
psycopg2-binary
whitenoise
beautifulsoup4
djangorestframework
djangorestframework-simplejwt
selenium
webdriver-manager
django-background-tasks